\name{internal.datasets}
\alias{medm9_proj}
\alias{regions.dim.bathy}
\docType{data}
\title{
internal datasets
}
\description{
internal (\code{lazyload}) datasets \code{medm9_proj} and \code{regions.dim.bathy}, accessed by \code{v.plot} and \link{readbin} respectively.
}
%\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
%}
%\references{
%%  ~~ possibly secondary sources and usages ~~
%}

\author{
Robert K. Bauer
}
