\name{oceanmap}
\alias{oceanmap}
\docType{package}
\title{
oceanmap - plot tools for 2D oceanographic data  
}
\description{
\code{oceanmap} is a plotting toolbox for oceanographic data.

\bold{Major functions:}\cr
\itemize{
\item{\link{plotmap}: plots landmask as basis or overlay}
\item{\link{v}: plots oceanographic data (fronts, SST, chla, bathymetry, etc.) from \link{raster}-objects, ncdf4- or gz-files}
\item{\link{set.colorbar}: adds colorbar to current figure, allowing several placement methods}
\item{\link{get.bathy}: download bahymetric data at user defined resolution from the NOAA ETOPO1 database}
\item{\link{add.region}: generate region definitions to facilitate landmask and colorbar plotting using \link{plotmap} and \link{v}}
\item{\link{figure} & \link{close_fig}: generate and save graphic devices in flexible fileformats}

}

\bold{Getting Started}\cr
Check out some examples of the principle functions, listed above.

}


\author{
Robert K. Bauer
}
